refactor: extract players tab list view from character panel
Agent-Logs-Url: https://github.com/wacky444/Zarka/sessions/175a313f-430b-4595-a65d-d5b587d5ce6e
